+++ CHANGES.rst
@@ -11,6 +11,17 @@ Unreleased
     some typing issues on cache control. :issue:`2881`
 
 
+Version 3.0.5
+-------------
+
+Unreleased
+
+-   The Watchdog reloader ignores file closed no write events. :issue:`2945`
+-   Logging works with client addresses containing an IPv6 scope :issue:`2952`
+-   Ignore invalid authorization parameters. :issue:`2955`
+-   Improve type annotation fore ``SharedDataMiddleware``. :issue:`2958`
+
+
 Version 3.0.4
 -------------